refactor: 서비스 인터페이스와 구현을 하나의 클래스로 통합
AuthServiceImpl 제거하고, AuthService 클래스에 통합
